Refactor axios config header
