+++ client/src/utilities/axios-helper.js
@@ -1,0 +1,7 @@
+const { getItem } = require("./cookie-helper");
+
+const getToken = name => {
+  return getItem(name);
+};
+
+export const reqConfig = { headers: { Authorization: `Bearer ${getToken("token")}` } };
