+++ client/src/utilities/cookie-helper.js
